+++ test/endpoints/baz/baz_metrics_test.go
@@ -78,7 +78,7 @@ func TestCallMetrics(t *testing.T) {
 	headers["device"] = "ios"
 	headers["deviceversion"] = "carbon"
 
-	numMetrics := 13
+	numMetrics := 14
 	cg.MetricsWaitGroup.Add(numMetrics)
 
 	_, err = gateway.MakeRequest(
@@ -168,7 +168,7 @@ func TestCallMetrics(t *testing.T) {
 	tchannelTags := map[string]string{
 		"env":             "test",
 		"app":             "test-gateway",
-		"service":         "test-gateway",
+		"service":         "bazService",
 		"target-service":  "bazService",
 		"target-endpoint": "SimpleService__call",
 		"host":            zanzibar.GetHostname(),